Moved admin api tests out from ceramic daemon tests to a separate file
